Move non-opencl stuff
